% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExampleMatchedDataSet.R
\name{makeExampleMatchedDataSet}
\alias{makeExampleMatchedDataSet}
\title{Function for generating an example matchRanges dataset}
\usage{
makeExampleMatchedDataSet(type = "GRanges")
}
\arguments{
\item{type}{Character designating which type of dataset to make.
options are one of 'data.frame', 'data.table', 'DataFrame',
'GRanges', or 'GInteractions'.}
}
\value{
Returns an example dataset for
input to \code{matchRanges()}.
}
\description{
This function will generate an example dataset as input
for \code{matchRanges()}.
}
\examples{
## Make examples for matchRanges()
makeExampleMatchedDataSet()
makeExampleMatchedDataSet(type = 'data.frame') |> head()
makeExampleMatchedDataSet(type = 'data.table')
makeExampleMatchedDataSet(type = 'DataFrame')
makeExampleMatchedDataSet(type = 'GRanges')

}
